Delete: chatbot.py && Modify: migrate old function
